+++ aed1/vector/include/vetor.h
@@ -4,7 +4,7 @@
 
 typedef struct Vector {
     int tam;
-    long* v;
+    int* v;
 } Vector;
 
 Vector* initialize_v(int tam);
@@ -13,7 +13,7 @@ void destroy_v(Vector* v);
 
 int get_v(Vector* v, int i);
 
-void set_v(Vector* v, int i, long value);
+void set_v(Vector* v, int i, int value);
 
 void randomize_values_v(Vector* v, int seed, int size);
 
@@ -21,11 +21,11 @@ void randomize_values_asc_v(Vector* v, int seed, int size);
 
 void show_portion_v(Vector* v, int inicio, int);
 
-void swap_v(long* a, long* b);
+void swap_v(int* a, int* b);
 
-int search_sequencial_v(Vector* v, long i);
+int search_sequencial_v(Vector* v, int i);
 
-int search_binario_v(Vector* v, long i);
+int search_binario_v(Vector* v, int i);
 
 void bubble_sort_v(Vector* v);
